rename sample to alarm
