+++ src/alarm/Controller/OnCompleteSettingAlarm.java
@@ -1,4 +1,4 @@
-package sample.Controller;
+package alarm.Controller;
 
 public interface OnCompleteSettingAlarm {
     void onComplete(int totalTime);
