+++ src/alarm/Main.java
@@ -1,4 +1,4 @@
-package sample;
+package alarm;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
