+++ src/alarm/Sudoku/SudokuController.java
@@ -1,4 +1,4 @@
-package sample.Sudoku;
+package alarm.Sudoku;
 
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
