+++ src/alarm/View/gameSelector.fxml
@@ -3,7 +3,7 @@
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.layout.GridPane?>
-<GridPane fx:controller="sample.Controller.GameSelectorController"
+<GridPane fx:controller="alarm.Controller.GameSelectorController"
           xmlns:fx="http://javafx.com/fxml"
           alignment="center" >
     <children>
