+++ src/alarm/View/setAlarmScreen.fxml
@@ -8,7 +8,7 @@
 
 
 
-<AnchorPane prefHeight="420.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="sample.Controller.SetAlarmScreenController">
+<AnchorPane prefHeight="420.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="alarm.Controller.SetAlarmScreenController">
 
    <children>
       <Label layoutX="171.0" layoutY="30.0"  textFill="#404040">
