+++ src/alarm/View/sudokuResultCheck.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="sample.Sudoku.SudokuController">
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="alarm.Sudoku.SudokuController">
     <children>
         <Button layoutX="284.0" layoutY="253.0" mnemonicParsing="false" text="ok" textAlignment="CENTER" />
         <Label fx:id="popupErrorList" layoutX="133.0" layoutY="111.0" prefHeight="89.0" prefWidth="309.0" textAlignment="CENTER" textOverrun="CENTER_ELLIPSIS">
